+++ CHANGELOG.md
@@ -7,6 +7,9 @@ and this project adheres to [Semantic Versioning](https://semver.org/spec/v2.0.0
 
 ## [Unreleased]
 
+### Added
+- `stop_task` control request and `Client#stop_task` method for stopping running background tasks (TypeScript SDK parity)
+
 ## [0.7.5] - 2026-02-10
 
 ### Added